pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "jota-dev00-0",email = "jpar.dev00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
